requirements.txt
asgiref==3.7.2
black==23.3.0
Brotli==1.0.9
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dj-database-url==2.0.0
Django==4.2.3
django-filter==23.2
django-getenv==1.3.2
djangorestframework==3.14.0
exceptiongroup==1.1.2
getenv==0.2.0
gunicorn==20.1.0
idna==3.4
iniconfig==2.0.0
Markdown==3.4.3
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.8.0
pluggy==1.2.0
psycopg==3.1.9
psycopg2-binary==2.9.6
pytest==7.4.0
pytest-django==4.5.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
responses==0.23.1
sqlparse==0.4.4
tomli==2.0.1
types-PyYAML==6.0.12.10
typing_extensions==4.7.1
urllib3==2.0.3
whitenoise==6.5.0
